split sorters up into multiple files
