+++ src/sort.rs
@@ -5,6 +5,9 @@ use egui::{
 
 use rand::{seq::SliceRandom, thread_rng};
 
+use crate::bubble_sort::BubbleSort;
+use crate::insertion_sort::InsertionSort;
+
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum SortType {
     BubbleSort,
@@ -35,72 +38,8 @@ pub(crate) trait Sorter {
 }
 
 #[derive(Default)]
-pub struct BubbleSort<T> {
-    current_vector: Vec<T>,
-    index: usize,
-    ceiling: usize,
-}
-
-#[derive(Default)]
 pub struct BogoSort<T> {
     current_vector: Vec<T>,
-}
-
-#[derive(Default)]
-pub struct InsertionSort<T> {
-    current_vector: Vec<T>,
-    index: usize,
-}
-
-impl<T> Sorter for BubbleSort<T>
-where
-    T: Clone + Copy + Ord,
-{
-    type ElementType = T;
-
-    fn next(&mut self) {
-        if self.ceiling == 0 {
-            return;
-        }
-        let mut vec = self.current_vector.clone();
-        let index = self.index;
-        if index < self.ceiling && vec[index] > vec[index + 1] {
-            vec.swap(index, index + 1)
-        }
-
-        self.index += 1;
-
-        if self.index >= self.ceiling {
-            self.ceiling -= 1;
-            self.index = 0;
-        }
-        self.current_vector = vec;
-    }
-
-    fn init(&mut self, v: Vec<T>) {
-        self.ceiling = v.len() - 1;
-        self.current_vector = v;
-    }
-
-    fn visualize(&mut self, ui: &mut PlotUi, convert: fn(Self::ElementType) -> f64) {
-        let mut v = self
-            .current_vector
-            .iter()
-            .enumerate()
-            .map(|(x, y)| Bar::new(x as f64 + 0.25, convert(*y) as f64))
-            .collect();
-
-        color(
-            &mut v,
-            vec![self.index, self.index + 1],
-            Color32::DARK_GRAY,
-            Color32::WHITE,
-        );
-
-        let chart = plot::BarChart::new(v);
-
-        ui.bar_chart(chart);
-    }
 }
 
 impl<T> Sorter for BogoSort<T>
@@ -134,71 +73,7 @@ where
     }
 }
 
-impl<T> InsertionSort<T>
-where
-    T: Ord,
-{
-    fn binary_search(&self, begin: usize, end: usize, item: &T) -> usize {
-        if begin >= end {
-            return if item > &self.current_vector[begin] {
-                begin + 1
-            } else {
-                begin
-            };
-        }
-
-        let mid = (begin + end) / 2;
-
-        if item > &self.current_vector[mid] {
-            return self.binary_search(mid + 1, end, item);
-        }
-
-        if item < &self.current_vector[mid] {
-            return self.binary_search(begin, if mid > 0 { mid - 1 } else { 0 }, item);
-        }
-
-        mid + 1
-    }
-}
-
-impl<T> Sorter for InsertionSort<T>
-where
-    T: Ord + Clone + Copy,
-{
-    type ElementType = T;
-
-    fn next(&mut self) {
-        if self.index >= self.current_vector.len() {
-            return;
-        }
-        let val = self.current_vector[self.index].clone();
-        let i = self.binary_search(0, self.index, &val);
-        self.current_vector.insert(i, val);
-        self.current_vector.remove(self.index + 1);
-        self.index += 1;
-    }
-
-    fn init(&mut self, v: Vec<Self::ElementType>) {
-        self.current_vector = v;
-    }
-
-    fn visualize(&mut self, ui: &mut PlotUi, convert: fn(Self::ElementType) -> f64) {
-        let mut v = self
-            .current_vector
-            .iter()
-            .enumerate()
-            .map(|(x, y)| Bar::new(x as f64 + 0.25, convert(*y) as f64))
-            .collect();
-
-        color(&mut v, vec![self.index], Color32::DARK_GRAY, Color32::WHITE);
-
-        let chart = plot::BarChart::new(v);
-
-        ui.bar_chart(chart);
-    }
-}
-
-fn color(v: &mut Vec<Bar>, special: Vec<usize>, base: Color32, highlight: Color32) {
+pub(crate) fn color(v: &mut Vec<Bar>, special: Vec<usize>, base: Color32, highlight: Color32) {
     for mut bar in v {
         let arg = bar.argument as usize;
         bar.fill = if special.contains(&arg) {
